interninfos/requirements.txt
flask
flask-jwt-extended
psycopg2-binary
python-dotenv
werkzeug
nltk
spacy
transformers
torch
pandas
xlsxwriter
reportlab
openpyxl
scikit-learn
langdetect
sentence-transformers
redis
rq
pytest
